pyproject.toml
[project]
name = "spark_fraud_detection"
version = "0.1.0"
description = "Fraud detection pipeline w/ Spark"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "awscli>=1.33.0",
    "boto3>=1.35.0",
    "duckdb>=1.3.2",
    "faker>=37.5.3",
    "findspark>=2.0.1",
    "grpcio>=1.74.0",
    "grpcio-status>=1.74.0",
    "ibis-framework[pyspark]>=10.8.0",
    "numpy>=2.3.2",
    "pandas>=2.3.1",
    "protobuf>=6.31.1",
    "pyspark==3.5.6",
]

[dependency-groups]
dev = [
    "marimo>=0.14.16",
]
